+++ test/test.js
@@ -1298,10 +1298,15 @@
         base: '/path/to/file?some=query#hash',
         result: '/path/relative/path?blubber=1#hash3'
       }, {
-        name: 'base query string',
+        name: 'fragment absolute url',
         url: '#hash3',
         base: '/path/to/file?some=query#hash',
         result: '/path/to/file?some=query#hash3'
+      }, {
+        name: 'fragment relative url',
+        url: '#hash3',
+        base: 'path/to/file',
+        result: 'path/to/file#hash3'
       }, {
         name: 'relative path - urljoin',
         url: 'the_relative_url',